+++ tests/metal/tuners/test_tuner.py
@@ -2,26 +2,27 @@ import unittest
 
 import numpy as np
 
-from metal.tuner import ModelTuner, HyperbandTuner
+from metal.tuners.random_tuner import RandomSearchModelTuner
+from metal.tuners.hyperband_tuner import HyperbandModelTuner
 
 class TunerTest(unittest.TestCase):
 
     def test_config_constant(self):
         search_space = {'a': 1}
-        tuner = ModelTuner(None, None, 123)
+        tuner = RandomSearchModelTuner(None, None, 123)
         configs = list(tuner.config_generator(search_space, max_search=10))
         self.assertEqual(len(configs), 1)
 
     def test_config_list(self):
         search_space = {'a': [1, 2]}
-        tuner = ModelTuner(None, None, 123)
+        tuner = RandomSearchModelTuner(None, None, 123)
         configs = list(tuner.config_generator(search_space, max_search=10))
         self.assertEqual(len(configs), 2)
 
     def test_config_two_values(self):
         search_space = {'a': [1],
                         'b': [1, 2, 3]}
-        tuner = ModelTuner(None, None, 123)
+        tuner = RandomSearchModelTuner(None, None, 123)
         configs = list(tuner.config_generator(search_space, max_search=10))
         self.assertEqual(len(configs), 3)
 
@@ -29,7 +30,7 @@ class TunerTest(unittest.TestCase):
         search_space = {'a': [1],
                         'b': [1, 2, 3],
                         'c': {'range': [1, 10]}}
-        tuner = ModelTuner(None, None, 123)                        
+        tuner = RandomSearchModelTuner(None, None, 123)                        
         configs = list(tuner.config_generator(search_space, max_search=10))
         self.assertEqual(len(configs), 10)
 
@@ -37,7 +38,7 @@ class TunerTest(unittest.TestCase):
         search_space = {'a': [1],
                         'b': [1, 2, 3],
                         'c': {'range': [1, 10]}}
-        tuner = ModelTuner(None, None, 123)                        
+        tuner = RandomSearchModelTuner(None, None, 123)                        
         configs = list(tuner.config_generator(search_space, max_search=0))
         self.assertEqual(len(configs), 3)
 
@@ -46,7 +47,7 @@ class TunerTest(unittest.TestCase):
                         'b': [1, 2, 3],
                         'c': {'range': [1, 10]},
                         'd': {'range': [1, 10], 'scale': 'log'}}
-        tuner = ModelTuner(None, None, 123)                        
+        tuner = RandomSearchModelTuner(None, None, 123)                        
         configs = list(tuner.config_generator(search_space, max_search=20))
         self.assertEqual(len(configs), 20)
         self.assertGreater(
@@ -58,7 +59,7 @@ class HyperbandTunerTest(unittest.TestCase):
     def test_hyperband_schedule_correctness(self):
         # Test the default schedule that is generated by the hyperband tuner
         # (which at the moment is budget=200, eta=3)
-        hyperband_tuner = HyperbandTuner(None,
+        hyperband_tuner = HyperbandModelTuner(None,
                                          hyperband_epochs_budget=200, 
                                          hyperband_proportion_discard=3, 
                                          seed=123)
@@ -74,7 +75,7 @@ class HyperbandTunerTest(unittest.TestCase):
     def test_hyperband_paper_schedule_correctness(self):
         # Generate the schedule in the hyperband paper
         # (https://arxiv.org/pdf/1603.06560.pdf)
-        hyperband_tuner = HyperbandTuner(None,
+        hyperband_tuner = HyperbandModelTuner(None,
                                          hyperband_epochs_budget=1701, 
                                          hyperband_proportion_discard=3, 
                                          seed=123)
